requirements.txt
Flask==2.3.3
spidev==3.6
RPi.GPIO==0.7.1